requirements.txt
# Core dependencies
kivy>=2.2.0
kivy[base,media]
kivy-deps.angle>=0.3.3
kivy-deps.glew>=0.3.1
kivy-deps.sdl2>=0.6.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=9.5.0
numpy>=1.24.0

# Google Vision API (optional)
google-cloud-vision>=3.4.0
google-auth>=2.17.0

# GPS and Location
requests>=2.31.0

# Data handling
psutil>=5.9.0

# Map visualization (optional)
folium>=0.14.0
webbrowser2>=0.4.1

kivy==2.2.0
google-cloud-vision==3.4.4
pillow==10.0.0
requests==2.31.0
numpy==1.24.3
opencv-python==4.8.1.78
plyer==2.1.0
